requirements.txt
aniso8601==3.0.0
astroid==1.6.3
autopep8==1.3.5
boto3==1.7.4
botocore==1.10.4
click==6.7
docutils==0.14
Flask==0.12.2
flask-restplus==0.10.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.4.0
pylint==1.8.4
python-dateutil==2.6.1
pytz==2018.4
s3transfer==0.1.13
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
pymongo==3.5.1
flask_sqlalchemy==2.3.1
flask-migrate==2.2.1
psycopg2==2.7.5
flask_script==2.0.6